renamed the zoo folder
